==> src/non_expert_annotations/02_non_expert_annotations_data_cleaning.Rmd <==
---
title: "01_data_cleaning"
author: "Jacob Beck"
date: "`r Sys.Date()`"
output: distill::distill_article
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
```

# Load packages

```{r message=FALSE, warning=FALSE}
library(tidyverse)
```

# Load and preprocess Data

-   Load data
-   Rename variables

```{r warning=FALSE}
raw_data <- read.csv("../../data/processed/non_expert_annotations/non_expert_annotations_raw.csv")

df <- raw_data %>% rename(
  page_used = Page.used,
  llm_scope = LLM.Year, # column year and scope swapped in original excel by mistake
  llm_value = LLM.value,
  llm_unit = LLM.unit,
  llm_year = LLM.Scope, # column year and scope swapped in original excel by mistake
  value_correct = Value.correct..Yes.No.,
  corrected_value = Value.corrected..fill.in.if.necessary.,
value_reasoning = Value.Reasoning..select.if.necessary..if.multiple.apply.choose.first.,
unit_correct = Unit.correct..Yes.No.,
corrected_unit = Unit.corrected..fill.in.if.necessary.,
    unit_reasoning = Unit.Reasoning..select.if.necessary..if.multiple.apply.choose.first.,
    page_correct = Page.correct..Yes.No.,
    corrected_page = Page.corrected..fill.in.if.necessary.,
    reporting_type = Reporting.Type..Table.Graphic..Text....from.PDF,
    metric_name = Emission.Metric.Name.from.PDF,
    comment = Comment..if.helpful.,
    record_expert_needed = Record.needs.expert.adjudication..Yes.No.,
    exp_corrected_value = Expert.Value.corrected,
    exp_corrected_unit = Expert.unit.corrected,
    exp_corrected_page = Expert.page.corrected,
    exp_comment = Expert.Comment,
    exp_problem = Expert.cannot.determine.correct.value..unit.or.page,
    PageTextsToLLM = page_texts_to_llm,
    TextResponseFromLLM = text_response_from_llm,
    report_name = Report.Name.,
    pages_searched = Pages.searched.,
    annotator_ID = Annotator.ID.,
    reporting_standards = Reporting.standards.,
    company_boundaries = Handling.company.boundaries.,
    document_expert_needed = Document.needs.expert.adjudication.,
    document_comment = Document.comment.,
    expert_ID = Expert.ID.
)
```

# Clean data

## Handle missing annotator ID

```{r}
df %>% filter(is.na(annotator_ID)) %>%
  select(report_name, annotator_ID) %>%
  distinct(report_name)

```

=\> 1 reports has a missing annotator ID

Manually retrieved annotator IDs from the original data:

| report_name   | annotator_ID |
|---------------|--------------|
| kitz 2019     | 8            |

Put in the corrected annotator IDs

```{r}
df1 <- df %>%
  mutate(
    annotator_ID = ifelse(report_name == "kitz corp_2019_report.pdf" & is.na(annotator_ID), 8, annotator_ID))

table(df1$annotator_ID) # check that there are more or less the same number of entries for every annotator
```

## Correct text columns
- Some columns contain the entry "n/a" which should be replaced by NA. We first check which columns are concerned and then correct it.
- Some columns which originally contained numeric and string values are loaded as string columns to avoid any data loss. As a result, a value like "110" is saved as "110.0". This concerns the columns ID, corrected_page and llm_scope. To correct this, we split off the ".0" from the entries.

```{r}
# Check which columns of df1 contain entry "n/a"
for (col in colnames(df1)) {
  if (any(df1[[col]] == "n/a", na.rm = TRUE)) {
    print(col)
  }  
} 

df2 <- df1 %>%
  mutate(
    # replace "n/a" with NA
    value_correct = ifelse(value_correct == "n/a", NA, value_correct),
    corrected_value = ifelse(corrected_value == "n/a", NA, corrected_value),
    unit_correct = ifelse(unit_correct == "n/a", NA, unit_correct),
    corrected_unit = ifelse(corrected_unit == "n/a", NA, corrected_unit),
    page_correct = ifelse(page_correct == "n/a", NA, page_correct),
    corrected_page = ifelse(corrected_page == "n/a", NA, corrected_page),
    reporting_type = ifelse(reporting_type == "n/a", NA, reporting_type),
    metric_name = ifelse(metric_name == "n/a", NA, metric_name),
  ) %>%
  mutate(
    # split off ".0" from ID
    ID = str_replace_all(ID, pattern = "\\b(\\d+)\\.0\\b", replacement = "\\1"),
    # split off ".0" from corrected_page
    corrected_page = str_replace_all(corrected_page, pattern = "\\b(\\d+)\\.0\\b", replacement = "\\1"),
    # split off ".0" from llm_scope
    llm_scope = str_replace_all(llm_scope, pattern = "\\b(\\d+)\\.0\\b", replacement = "\\1"),
  )

```


## Handle the rows with row ID "x"

The ID column contains entries of numeric form "112" as well as of string form "x". The latter concerns rows which were added manually by the non-experts. In order to create a unique identifier for these rows, we concatenate their ID with corrected_page, llm_year and llm_scope.

```{r}
df3 <- df2 %>%
  mutate(ID = ifelse(ID == "x",paste0("x_", corrected_page,"_", llm_year,"_", llm_scope),ID))
```

## Handle wrong formatting in corrected_value  

Inspect all the values
```{r}
df3 %>% select(corrected_value) %>% distinct()
```

While llm_value is a numeric column, corrected_value is a character column. We want to transform corrected_value to numeric, but this might create problems if values inserted by annotators have wrong format.

- Annotators sometimes use ',' in 6-digit numbers to separate thousands ==> remove ','
- Annotators sometimes use '.' in 6-digit numbers to separate thousands ==> remove '.'
- value "209,034.43 4" seems to be a typing error ==> remove last character
- Instead of leaving the cell empty, the annotators sometimes wrote "-" or n/a ==> set to ""

```{r}
pattern <- "^\\d{3}\\.\\d{3}$" # e.g. 209.034
df4 <- df3 %>% mutate(corrected_value = str_replace_all(corrected_value, "[-,]", ""),
                      corrected_value = ifelse(corrected_value == "209034.43 4", "209034.43", corrected_value),
                      corrected_value = str_replace_all(corrected_value, pattern, str_replace_all(corrected_value, "\\.", "")),)
```

Now try conversion to numeric again
```{r}
df4$corrected_value <- as.numeric(df4$corrected_value)
```

## Handle falsely set reporting_type

If the value_correct is "No" and corrected_value is NA, the reporting_type and metric_name should be NA because the latter refer to the true value. If the true value is NA, reporting type and metric must be NA by definition.

```{r}
df4 %>% filter(value_correct == "No" & is.na(corrected_value) & (!is.na(reporting_type) | !is.na(metric_name)))
```

```{r}
df5 <- df4 %>% mutate(
  reporting_type = ifelse(value_correct == "No" & is.na(corrected_value), NA, reporting_type), 
  metric_name = ifelse(value_correct == "No" & is.na(corrected_value), NA, metric_name))
```

## Handle annotation mistakes
If value_correct == "Yes", corrected_value should be NA
If corrected_value is not NA, value_correct should be "No"
Equivalently for unit and page

For value:
- Is value_correct == NA & corrected_value not NA?

```{r}
df5%>% filter(is.na(value_correct) & !is.na(corrected_value))

```
Set value_correct to "No"
```{r}
df6 <- df5 %>% mutate(value_correct = ifelse(is.na(value_correct) & !is.na(corrected_value), "No", value_correct))
```

- Is value correct == "Yes" & corrected_value not NA?
```{r}
df6 %>% filter(value_correct == "Yes" & !is.na(corrected_value))
```

For unit:
- Is value_correct == NA & corrected_value not NA?
```{r}
df6 %>% filter(is.na(unit_correct) & !is.na(corrected_unit))
```
```{r}
df7 <- df6 %>% mutate(unit_correct = ifelse(is.na(unit_correct) & !is.na(corrected_unit), "No", unit_correct))
```

- Is unit_correct == "Yes" & corrected_unit not NA?
```{r}
df7 %>% filter(unit_correct == "Yes" & !is.na(corrected_unit))
```

For page:
- Is page_correct NA & corrected_page not NA?
```{r}
df7 %>% filter(is.na(page_correct) & !is.na(corrected_page))
```

- Is page_correct == "Yes" & corrected_page not NA?
```{r}
df7 %>% filter(page_correct == "Yes" & !is.na(corrected_page))
```
Correct this manually
```{r}
df8 <- df7 %>% 
  mutate(page_correct = ifelse(page_correct == "Yes" & !is.na(corrected_page), "No", page_correct))
```

# Save cleaned data
```{r}
write.csv(df8,"../../data/processed/non_expert_annotations/non_expert_annotations_cleaned.csv", row.names = FALSE)
```
